Moves the selectPrintCommand function from the simple module to the inputs module
